requirements.txt
requests
twilio
python-dotenv
schedule
pycountry
phonenumbers
tkcalendar
pytz
pillow
pystray
tabulate
coverage
pytest
pytest-cov